requirements.txt
torch
torchvision
matplotlib
trimesh
pygltflib
numpy
seaborn
gradio
pillow
